% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_migramod.R
\name{fit_migramod}
\alias{fit_migramod}
\title{Estimate a migramod model}
\usage{
fit_migramod(dataIn = dataIn, paramaters_0, model.rc)
}
\arguments{
\item{dataIn}{Empiral for estimate}

\item{paramaters_0}{set of Paramater for Rogers castro function}

\item{model.rc}{Object of class migramod with roger Castro}
}
\description{
Estimate a migramod model via no linear estimation
}
