% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genrandom.R
\name{genRandomPar}
\alias{genRandomPar}
\title{Generator function for initial values in Rogers and Castro models}
\usage{
genRandomPar(profile = "seven")
}
\arguments{
\item{profile}{Number of parameters for the model that will be optimized.}
}
\value{
A list with random values for parameters
}
\description{
Generate 7, 9, 11 and 13 random values for Roger and Castro migration models.
}
\examples{
\dontrun{
genRandomPar(profile = "seven")
genRandomPar(profile = "nine")
genRandomPar(profile = "eleven")
genRandomPar(profile = "thirteen")
}


}
