% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migramod.R
\docType{class}
\name{MigraModel-class}
\alias{MigraModel-class}
\alias{MigraModel}
\title{MigraModel Reference Class}
\description{
The class incorporates the methods as the calculation of the Jacobian, the gradient and the Hessian matrix. This class is used to evaluate a given rc_expression with an initial tupla of values.
}
\details{
A Reference Class for fitting MigraModels
}
\section{Fields}{

\describe{
\item{\code{name}}{A character name}

\item{\code{expr}}{An rc_expression}
}}


