% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migraR.R
\docType{package}
\name{migraR}
\alias{migraR}
\alias{migraR-package}
\title{migraR: A package for computating models for migration analysis .}
\description{
The migraR package provides functions for estimate the parameters for Rogers and Castro Models:
\code{rc.7}, \code{rc.11}, \code{rc.13} like a package \code{migest}.
}
\section{migraR functions}{

The migraR functions were concived for estimate via non-linear optimization, the parameters of the Rogers and Castro
models wit seven, eleven and thirteen parameters. Besides, the package gives the user the posibility to investigate 
wich curve fits better using the method provided in the IUSSP manual for the estimation of Rogers and Castro multi 
exponential model migration schedule.
(http://demographicestimation.iussp.org/content/multi-exponential-model-migration-schedule).
}

