% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_models.R
\name{plot_RC}
\alias{plot_RC}
\title{Attempt for a function to make the graphics}
\usage{
plot_RC(dataIn)
}
\arguments{
\item{dataIn}{data for optimization}
}
\description{
Attempt for a function to make the graphics
}
\examples{
\dontrun{
plot_RC(dataIn)
}

}
