% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_expression.R
\name{rc_expression}
\alias{rc_expression}
\title{Rogers and Castro expressions (Rogers and Castro, 1981)}
\usage{
rc_expression(profile = "eleven")
}
\arguments{
\item{model.name}{Character with the the of migration model}

\item{An}{integer with the number of parameters in the model}
}
\description{
Rogers and Castro Expressions for migration by individual or grouped ages with methods
}
