% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_migramod.R
\name{best_migramod}
\alias{best_migramod}
\title{select best migramod model}
\usage{
best_migramod(dataIn = dataIn, model.rc, iter = 100, profile = "eleven")
}
\arguments{
\item{dataIn}{Set data for fit}

\item{model.rc}{a Object of class MigraMod}

\item{iter}{number of iterations for model}
}
\value{
a list with named parameters
}
\description{
One solution to initial values problems
}
\examples{
data("es_asmr")
data1 <- es_asmr[,c(1,5)]
colnames(data1) <- c("x","y")
model.rc.13 = MigraModel(
  name = 'castro_11',
  expr = rc_expression(profile = "thirteen")
)

model.rc7 = MigraModel(
  name = 'castro_7',
  expr = rc_expression(profile = "seven")
)
model.rc.11 = MigraModel(
  name = 'castro_11',
  expr = rc_expression(profile = "eleven")
)

plot(data1, cex=0.1, xlab = 'Age', ylab = 'Standarized Migration Rate')

fitted.val.11 <- best_migramod(dataIn = data1, model.rc =model.rc.11, iter = 50, profile = "eleven")
fitted.val.13 <- best_migramod(dataIn = data1, model.rc =model.rc.13, iter = 50, profile = "thirteen")
lines(data1[,1], model.rc7$value(fitted.val.7$bestParam,data1), col="red")
lines(data1[,1], model.rc.11$value(fitted.val.11$bestParam,data1), col="blue")
lines(data1[,1], model.rc.13$value(fitted.val.13$bestParam,data1), col="gold")
legend("topright",legend = c("seven","eleven","thirteen"),fill = c("red","blue","gold"))


}
