% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_migramod.R
\name{best_migramod}
\alias{best_migramod}
\title{Select best migramod model.}
\usage{
best_migramod(dataIn, profile = "eleven",
      maxite = 100, epsilon = 1e-05, datasimul = TRUE)
}
\arguments{
\item{dataIn}{Standarized migration data set for optimization.}

\item{profile}{Number of parameters of a Roger and Castro model.}

\item{maxite}{Maximum number of iterations for model optimization.}

\item{epsilon}{Tolerance in which the difference between the Mean Squared Error that will finish the algorithm.}

\item{datasimul}{Table with the different simulations and values of the parameters estimated.}
}
\value{
a list with named parameters
}
\description{
One solution to initial values problem based on the generaton of a random sample with size equals one from uniform apriori
distributions with values between 0 and 1 for the initial parameters.
}
\examples{
\dontrun{
library(migraR)
library(dplyr)
data("es_asmr")
data1 <- es_asmr[-c(1,2),c(1,5)]
colnames(data1) <- c("x","y")
attach(data1)

# Fitting and Plotting data
fitted.val.7 <- best_migramod(dataIn = data1,
                maxite = 5E2,
                profile = "seven")
fitted.val.9 <- best_migramod(dataIn = data1,
                maxite = 5E2,
                profile = "nine")
fitted.val.11 <- best_migramod(dataIn = data1,
                 maxite = 5E2,
                 profile = "eleven")
fitted.val.13 <- best_migramod(dataIn = data1,
                 maxite = 5E2,
                 profile = "thirteen")

x11()
plot(data1, cex=0.1, xlab = 'Age',
     ylab = 'Standarized Migration Rate')
lines(data1[,1],
      model.rc.7$value(fitted.val.7$bestParam,data1),
      col="blue")
lines(data1[,1],
      model.rc.9$value(fitted.val.9$bestParam,data1),
      col="orange")
lines(data1[,1],
      model.rc.11$value(fitted.val.11$bestParam,data1),
      col="blue", lty=3)
lines(data1[,1],
      model.rc.13$value(fitted.val.13$bestParam,data1),
      col="green")
legend('topright',
      legend = c(paste("(7)", "MAPE:", round(as.numeric(fitted.val.7$bestMAPE),2),
                 "R²:", round(as.numeric(fitted.val.7$bestRcuad),3)),
                 paste("(9)", "MAPE:", round(as.numeric(fitted.val.9$bestMAPE),2),
                 "R²:", round(as.numeric(fitted.val.9$bestRcuad),3)),
                 paste("(11)", "MAPE:", round(as.numeric(fitted.val.11$bestMAPE),2),
                 "R²:", round(as.numeric(fitted.val.11$bestRcuad),3)),
                 paste("(13)", "MAPE:", round(as.numeric(fitted.val.13$bestMAPE),2),
                 "R²:", round(as.numeric(fitted.val.13$bestRcuad),3))),
                 col = c("red",'orange',"blue","darkgreen"), lty = c(2,6,3,5))
}
}
