% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_asmr.R
\docType{data}
\name{es_asmr}
\alias{es_asmr}
\title{Spanish age specific migration rate data}
\format{A \code{data.frame} with 7 columns as follows:
\describe{
  \item{Age}{Specific age group}
  \item{male.foreign}{Age specific out migration rates abroad Spain for male}
  \item{female.foreign}{Age specific out migration rates abroad Spain for female}
}}
\usage{
es_asmr
}
\description{
Internal and International migration rates by sigle age and sex for Spain.
Rates calculated with using the data (migrants and population) form the Spanish Statistical Institute (INE).
}
\keyword{Migration}
\keyword{Spanish}
