% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_asmr.R
\docType{data}
\name{es_asmr}
\alias{es_asmr}
\title{Spanish age specific migration rate data}
\format{
A \code{data.frame} with 7 columns as follows:
\describe{
  \item{Age}{Specific age group}
  \item{male.foreign}{Age specific out migration rates abroad Spain for male}
  \item{female.foreign}{Age specific out migration rates abroad Spain for female}
  \item{female.foreign}{Age specific out migration rates abroad Spain for female}
  \item{male.inter_regional}{Age specific out migration rates abroad Spain for male}
  \item{female.inter_regional}{Age specific out migration rates abroad Spain for female}
 \item{male.intra_province}{Age specific out migration rates abroad Spain for male}
 \item{female.intra_province}{Age specific out migration rates abroad Spain for female}

}
}
\usage{
es_asmr
}
\description{
Internal and International migration rates by single age and sex for Spain.
Rates calculated with using the data (migrants and population) form the Spanish Statistical Institute (INE).
}
\keyword{Internal}
\keyword{International}
\keyword{Migration}
\keyword{Spanish}
