% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_migramod.R
\name{fit_migramod}
\alias{fit_migramod}
\title{Estimate a migramod model.}
\usage{
fit_migramod(data_In = dataIn, parameters_0, model.rc)
}
\arguments{
\item{model.rc}{Object of class migramod with usually a Rogers and Castro model expression.}

\item{dataIn}{Empirical (out/in) migration rates for estimation.}

\item{paramaters_0}{Set of initial parameters for selected Rogers and Castro model equation.}
}
\description{
Estimate a migramod model via linear optimization. The function takes out the tuples that
does not have gradient.

\examples{
require(migraR)
require(dplyr)
data("es_asmr")
data1 <- es_asmr[-c(1,2),c(1,5)]
colnames(data1) <- c("x","y") #this must be done before using fit_migramod
attach(data1) #this must be done before using fit_migramod
fit_migramod(data1, parameters_0, model.rc.7)
}

}
