% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_migramod.R
\name{fit_migramod}
\alias{fit_migramod}
\title{Estimate a MigraModel class object.}

\description{
Estimate a MigraModelObject-class model via linear optimization.
}

\usage{
fit_migramod(dataIn = dataIn, parameters_0, model.rc)
}

\arguments{
\item{dataIn}{Empirical (out/in) migration rates for estimation.}

\item{paramaters_0}{Set of initial parameters for selected Rogers and Castro model equation.}

\item{model.rc}{Object of class migramod with usually a Rogers and Castro model expression.}
}

\example{
See complete example in best_migramod function.
}
