% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_migramod.R
\name{fit_migramod}
\alias{fit_migramod}
\title{Estimate a migramod model.}
\usage{
fit_migramod(dataIn = dataIn, parameters_0, model.rc)
}
\arguments{
\item{dataIn}{Empirical (out/in) migration rates for estimation.}

\item{model.rc}{Object of class migramod with usually a Rogers and Castro model expression.}

\item{paramaters_0}{Set of initial parameters for selected Rogers and Castro model equation.}
}
\description{
Estimate a migramod model via linear optimization
}
