% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_migramod.R
\name{fit_migramod}
\alias{fit_migramod}
\title{Estimate a migramod model via linear optimization}
\usage{
fit_migramod (dataIn, parameters_0, model.rc)
}
\arguments{
\item{dataIn}{Empirical (out/in) migration rates for estimation.}

\item{parameters_0}{Set of initial parameters for selected Rogers and Castro model equation.}

\item{model.rc}{Object of class migramod with usually a Rogers and Castro model expression.}
}
\value{
A list of values calculated in the minimization as OptimResult, R squared or MAPE.
}
\description{
Estimate a migramod model via linear optimization
}
\examples{
\dontrun{
param0 <- genRandomPar(profile='seven')
name = 'Parametric_model',
expr = rc_expression(profile = 'seven') ,
dataIn = dataIn
)
fit_migramod(dataIn, param0, model.rc)
}


}
