% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migramod.R
\docType{class}
\name{MigraModelObject-class}
\alias{MigraModelObject-class}
\alias{MigraModel}
\title{Migration Rate Models Class}
\description{
Class with Rogers and castro Models.
}
\details{
The Rogesr and Castro models for a migration are so useful
}

