% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migramod.R
\docType{class}
\name{MigraModelObject-class}
\alias{MigraModelObject-class}
\alias{MigraModel}
\title{Migration Rate Models Class.}
\description{
Class built based on Rogers and castro Models.
}
\details{
The migramodel class is necessary to construct the optimization. The class incorpores
the methods as the calculation of the Jacobian, the gradient and the Hessian matrix.
This class is used to evaluate a given rc_expression with an initial tupla of values.
}

