% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migramod.R
\docType{class}
\name{MigraModelObject-class}
\alias{MigraModelObject-class}
\alias{MigraModel}
\title{Migration Rate Models Class.}
\description{
Class built based on Rogers and castro Models.
}
\details{

The MigraModelObjetct-class is the function that allows to create a new class of models. There are already four types (7, 9, 11 and 13 parameters). If a new model (p.e. with the student peak) need to be creted this function helps to set the model as class with its attributes (i.e. return simulation results and function values).

}

