% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migramod.R
\docType{class}
\name{MigraModelObject-class}
\alias{MigraModelObject-class}
\alias{MigraModel}
\title{Migration Rate Models Class.}
\arguments{
\item{name}{Name to the model that will be optimized}

\item{expr}{Rogers and Castro expression comming from rc_expression}
}
\description{
The class incorpores the methods as the calculation of the Jacobian, the gradient and the Hessian matrix.
This class is used to evaluate a given rc_expression with an initial tupla of values.
}
\section{Slots}{

\describe{
\item{\code{Migramodel}}{The migramodel class is necessary to construct the optimization.}
}}


