% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migraR.R
\docType{package}
\name{migraR}
\alias{migraR}
\title{migraR: A package for computing models for migration analysis .}
\description{
The migraR package provides functions for estimate the parameters for Rogers and Castro Models:
\code{rc.7}, \code{rc.9}, \code{rc.11}, \code{rc.13}.
}
\section{migraR functions}{

The migraR functions were conceived for estimate via linear optimization, the parameters of the Rogers and Castro
models wit seven, eleven and thirteen parameters. Besides, the package gives the user the possibility to investigate
which curve fits better using the method provided in the IUSSP manual for the estimation of Rogers and Castro multi
exponential model migration schedule.


(http://demographicestimation.iussp.org/content/multi-exponential-model-migration-schedule).
}

