% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migraR.R
\docType{package}
\name{migraR}
\alias{migraR}
\alias{migraR-package}
\title{migraR: A package for computating models for migration analysis .}
\description{
The migraR package provides functions for estimate the parameters for Rogers y Castro Models:
\code{rc.7}, \code{rc.9}, \code{rc.15} like a package \code{migest}.
}
\section{migraR functions}{

The migraR functions ...
}

