% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_expression.R
\name{rc_expression}
\alias{rc_expression}
\title{Rogers and Castro expressions}
\usage{
rc_expression(model.name = "RC.11", profile = 11, ...)
}
\arguments{
\item{model.name}{character with the the of migration model}

\item{profile}{a integer with the number of parameter for the model}
}
\description{
Are the Rogers and Castro expressions for  migration by individual or grouped ages with methods
}
