% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_expression.R
\name{rc_expression}
\alias{rc_expression}
\title{Rogers and Castro expressions (Rogers and Castro, 1981)}
\usage{
rc_expression(profile = "eleven")
}
\arguments{
\item{profile}{The schedule specification, 'profile' could be seven, nine, eleven or thirteen}
}
\description{
Rogers and Castro Expressions for migration schedules.
}
