% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_expression.R
\name{rc_expression}
\alias{rc_expression}
\title{Rogers and Castro expressions for migration schedules(Rogers and Castro, 1981)
The schedule specification, 'profile' could be seven, nine, eleven or thirteen.
The function can be used to be modified to build the own profile.}
\usage{
rc_expression(profile = "eleven")
}
\arguments{
\item{profile}{Expression that fix the number of parameters has to have the equation to optimize}
}
\description{
Rogers and Castro expressions for migration schedules(Rogers and Castro, 1981)
The schedule specification, 'profile' could be seven, nine, eleven or thirteen.
The function can be used to be modified to build the own profile.
}
